Remove `state export recipe` and its support routines.
