+++ pkg/platform/model/buildplanner.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"os"
 	"regexp"
+	"runtime"
 	"strconv"
 	"strings"
 	"time"
@@ -72,6 +73,28 @@ func logRequestVariables(req gqlclient.Request) {
 		}
 		logging.Debug("Build Expression: %s", string(beData))
 	}
+}
+
+type SolverError struct {
+	wrapped          error
+	validationErrors []string
+	isTransient      bool
+}
+
+func (e *SolverError) Error() string {
+	return "buildplan_error"
+}
+
+func (e *SolverError) Unwrap() error {
+	return e.wrapped
+}
+
+func (e *SolverError) ValidationErrors() []string {
+	return e.validationErrors
+}
+
+func (e *SolverError) IsTransient() bool {
+	return e.isTransient
 }
 
 func init() {
@@ -599,4 +622,19 @@ func VersionStringToRequirements(version string) ([]bpModel.VersionRequirement, 
 		})
 	}
 	return requirements, nil
+}
+
+func FilterCurrentPlatform(hostPlatform string, platforms []strfmt.UUID, cfg Configurable) (strfmt.UUID, error) {
+	platformIDs, err := filterPlatformIDs(hostPlatform, runtime.GOARCH, platforms, cfg)
+	if err != nil {
+		return "", errs.Wrap(err, "filterPlatformIDs failed")
+	}
+
+	if len(platformIDs) == 0 {
+		return "", locale.NewInputError("err_recipe_no_platform")
+	} else if len(platformIDs) > 1 {
+		logging.Debug("Received multiple platform IDs. Picking the first one: %s", platformIDs[0])
+	}
+
+	return platformIDs[0], nil
 }
